Rename RGB pixel files to match our convention
