+++ src/include/pixels/RGBPixel.hpp
